Removed class definitions from main.rb and added them to user.rb and boar.rb in lib directory.
