+++ lib/board.rb
@@ -1,0 +1,34 @@
+class Board
+  attr_reader :symb, :counter, :status, :b, :symbols
+
+  def initialize
+    @b = Array.new(3) { |i| Array.new(3) { |j| ((i * 3 + j + 1)).to_s } }
+    @counter = 9
+    @symb = ''
+    @symbols=%w[X O]
+  end
+
+  def turn(symb, loc)
+    if !loc.to_s.match?(/[1-9]/) || loc.to_s.length != 1 || @b[(loc - 1) / 3][(loc - 1) % 3].match?(/[XO]/)
+      @status = false
+    else
+      @status = true
+      @b[(loc - 1) / 3][(loc - 1) % 3] = symb
+      if winner(@b)
+        @counter = 0
+        @symb = symb
+      end
+      @counter -= 1
+    end
+  end
+
+  private
+
+  def winner(test)
+    trans = test.transpose
+    (0...test.size).collect { |i| test[i][i] }.uniq.count == 1 ||
+      (0...test.size).collect { |i| test[test.size - i - 1][i] }.uniq.count == 1 ||
+      test.map { |x| x.uniq.count }.include?(1) ||
+      trans.map { |x| x.uniq.count }.include?(1)
+  end
+end
