+++ lib/user.rb
@@ -1,0 +1,7 @@
+class User
+  attr_reader :symb, :name
+  def initialize(symb, name)
+    @name = name
+    @symb = symb
+  end
+end
